drift_detection/docker_drift_detection/requirements.txt
absl-py==1.4.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
calmjs.parse==1.3.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.2.1
executing==2.0.1
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.6.0
google-auth==2.31.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
hls4ml==0.8.1
humanfriendly==10.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.14.0
keras-tuner==1.4.7
keras2onnx==1.7.0
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.1
onnxconverter-common==1.14.0
onnxruntime==1.18.1
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parse==1.6.5
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
ply==3.11
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyDigitalWaveTools==1.1
Pygments==2.18.0
pyparser==1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
QKeras==0.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.0
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-model-optimization==0.8.0
termcolor==2.4.0
tf2onnx==1.16.1
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
